dz_4/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
aiosqlite
pydantic-settings
python-multipart
passlib[bcrypt]
python-jose[cryptography]
minio